Switch to apy-api SSI
