+++ settlement-distributions/bid-distribution/src/bin/cli.rs
@@ -1,3 +1,4 @@
+use bid_distribution::apy_api::fetch_ssr_pmpe;
 use bid_distribution::generators::bidding::generate_bid_settlements;
 use bid_distribution::generators::psr_events::generate_psr_settlements;
 use bid_distribution::generators::sam_penalties::generate_penalty_settlements;
@@ -5,8 +6,6 @@ use bid_distribution::rewards::load_rewards_from_directory;
 use bid_distribution::sam_meta::ValidatorSamMeta;
 use bid_distribution::settlement_config::BidDistributionConfig;
 use env_logger::{Builder, Env};
-use rust_decimal::Decimal;
-use serde::Deserialize;
 use settlement_common::protected_events::generate_protected_event_collection;
 use settlement_common::revenue_expectation_meta::RevenueExpectationMetaCollection;
 use settlement_common::settlement_collection::SettlementCollection;
@@ -66,28 +65,6 @@ struct Args {
     /// Base URL of the apy-api service (used to fetch SSR pmpe for the scoring epoch)
     #[arg(long, env, default_value = "https://apy.marinade.finance")]
     apy_api_url: String,
-}
-
-#[derive(Deserialize)]
-struct EpochPmpeEntry {
-    epoch: u64,
-    pmpe: Decimal,
-}
-
-#[derive(Deserialize)]
-struct EpochPmpeResponse {
-    epochs: Vec<EpochPmpeEntry>,
-}
-
-fn fetch_ssi_pmpe(apy_api_url: &str, epoch: u64) -> anyhow::Result<Decimal> {
-    let url = format!("{}/v1/epoch-pmpe/ssr", apy_api_url.trim_end_matches('/'));
-    info!("Fetching SSR pmpe for epoch {epoch} from {url}");
-    let resp: EpochPmpeResponse = reqwest::blocking::get(&url)?.error_for_status()?.json()?;
-    resp.epochs
-        .iter()
-        .find(|e| e.epoch == epoch)
-        .map(|e| e.pmpe)
-        .ok_or_else(|| anyhow::anyhow!("SSR for epoch {epoch} not yet available at {url}"))
 }
 
 fn main() -> anyhow::Result<()> {
@@ -185,12 +162,7 @@ fn main() -> anyhow::Result<()> {
             rewards_collection.total_rewards()
         );
 
-        anyhow::ensure!(
-            !sam_validator_metas.is_empty(),
-            "SAM meta collection is empty"
-        );
-
-        let ssi_pmpe = fetch_ssi_pmpe(&args.apy_api_url, stake_meta_epoch)?;
+        let ssi_pmpe = fetch_ssr_pmpe(&args.apy_api_url, stake_meta_epoch)?;
         info!("SSI: {ssi_pmpe} pmpe (from apy-api, epoch {stake_meta_epoch})");
 
         // Epoch consistency verification
@@ -217,7 +189,7 @@ fn main() -> anyhow::Result<()> {
             bidding_config,
             &bid_distribution_config.fee_config,
             &*stake_authority_filter,
-            Some(ssi_pmpe),
+            ssi_pmpe,
         )?;
         info!("Generated {} bid settlements", bid_settlements.len());
         all_settlements.extend(bid_settlements);
