+++ settlement-distributions/bid-distribution/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod apy_api;
 pub mod generators;
 pub mod rewards;
 pub mod sam_meta;
